[chore] employee의 department 칼럼 변화에 따른 수정(entity, dto, 다른 메소드에서 쓰인 builder 부분 등) 다른 팀원들도 코드 수정 요망

+++ src/main/java/kr/co/groupworks/dto/cis/employee/DepartmentDTO.java
@@ -13,7 +13,7 @@ import lombok.experimental.Accessors;
 @AllArgsConstructor
 public class DepartmentDTO {
     // 1, 부서 pk
-    private int departmentId;
+    private Long departmentId;
     // 2, 부서 명
     private String departmentName;
     // 3, 부서 공용 연락처
